+++ CHANGELOG.md
@@ -3,11 +3,13 @@ All notable changes to this project will be documented in this file.
 
 ## master branch
 ### Added
+- class `DroitHistory` added to `droit.models`
 - class `DroitCache` added to `droit.models`
 - telegram io module added
 
 ### Changed
-- class `DroitResourcePackage` has new attribute cache which contains a `DroitCache` object
+- class `DroitResourcePackage` has new attribute `hisotry` which contains a `DroitHistory` object
+- class `DroitResourcePackage` has new attribute `cache` which contains a `DroitCache` object
 - [#17](https://github.com/jaybeejs/python-droit/issues/17) input plugins return rpack
 
 ## 1.0.0 pre1